refactor register controller and service into auth
